background-service/service/pyproject.toml
[tool.poetry]
name = "background-app"
version = "0.1.0"
description = ""
authors = ["Michael <mihail.tchernov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.12"
uvicorn = "^0.27.0.post1"
fastapi = "^0.109.2"
pydantic = "^1.10.7"
poetry = "^1.7.1"
gstream = {path = "./dist/gstream-0.1.0-py3-none-any.whl"}
python-multipart = "^0.0.9"
redis = "^5.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
